[Chore] 스팟 조회 api 수정
좋아요 수, 방명록 수 등 추가

+++ src/main/java/com/ttubeog/domain/spot/dto/request/UpdateSpotRequestDto.java
@@ -25,10 +25,10 @@ public class UpdateSpotRequestDto {
     private String info;
 
     @Schema(description = "위도", defaultValue = "36.5642886")
-    private Float latitude;
+    private Double latitude;
 
     @Schema(description = "경도", defaultValue = "126.9275718")
-    private Float longitude;
+    private Double longitude;
 
     @Schema(description = "이미지 리스트")
     private List<String> image;
